Specify system server compiler filter if configured
Not all devices have dalvik.vm.systemservercompilerfilter defined, but
when they do, we need to respect the vendor's configuration.

ComposHostTestCases now tests both speed and speed-profile explicitly,
rather than using the default value (which should be tested later in the
e2e test).

Add services.jar.prof to the temporary allowlist so that it can be
accessed from the VM's ANDROID_ROOT.

Bug: 208269838
Test: atest ComposHostTestCases
Test: setprop dalvik.vm.systemservercompilerfilter speed-profile
      composd_cmd async-odrefresh
      # Observed the odrefresh flag in VM's log
Change-Id: I1348a62834b9a65dde3e74a648394916473cda16