Allow CompOS VM to use all CPUs by default
In early boot, if dalvik.vm.boot-dex2oat-threads is not specified,
dex2oat will run in #CPU threads.

Currently, we're still using one CPU in the VM if
dalvik.vm.boot-dex2oat-threads not specified. This change gives the VM
the same number of CPUs, which will allow dex2oat to leverage all CPUs.

Bug: 197358423
Test: Run `composd_cmd staged-apex-compile` on oriole
      # Before: ~90s
      # After:  ~60s
Test: enable adb, check /proc/cpuinfo
Change-Id: Ifb8908ab44eacf42960f6e4d9b12edca6370d594