Reduce vm timeout
We increased the timeout during the switch to protected
(commit bf209f8c842975fc94240ffd9b7576fe90b3457a), but partly because
of various bugs that were causing slowness. Reduce it again, but back
to 15s instead of the original 10s just to be cautious.

While I'm here, delete a now-unused timeout from the
odrefresh-in-the-host days.

Bug: 220824234
Bug: 218494522
Test: Presubmit
Change-Id: Idba2c5abae500ecc26b4ca731a32680dab58b43e