Orderly VM shutdown, part 1
Rather than killing the VM when we are done with it, ask the payload
to exit, and wait for the VM to have full exited. This allows the VM
time to write logs, generated crash dumps etc if there has been a
failure.

Add a quit method to the CompOS service, so clients can request it to
exit. Add the ability to wait for a VM to have died with a timeout to
vmclient. Implement a wait for shutdown in compos_client that waits
for the VM to exit but terminates it abruptly if it doesn't do so in a
reasonable time; do the same thing if the VM fails to start.

Change compos_verify to use this method to wait for the VM to have
fully exited once we are done with it.

Assorted refactorings:

- Simplify the timeout handling code so we panic if the neccessary
  property isn't available (all requests would fail anyway). Also
  updated the timeouts a little.

- Rename get_service to connect_service (it's definitely not a simple
  getter).

I haven't dealt with compilation yet; that will have ramifications all
the way up to Java, and this CL is big enough already. Additionally I
haven't yet attempted to allow odsign to continue while we wait for
the VM to exit.

Bug: 236581575
Test: Run VM, see both finished & died in the logs.
Change-Id: I47501081d23833fe7ef791240161d93e38b3d951