Clean up instance tracking
We used to hold a Weak<CompOsInstance>, so that we could allow clients
to retrieve the current running instance. But we haven't supported
that since commit 7d040d413411b6c0da06a00ae6f470b3a84619ca, and it
confuses the ownership semantics.

We still need to track whether an instance is running, but we can do
that via a separate Arc<()> that is kept alive by the instance.

Bug: 236581575
Test: compos_cmd test-compile
Test: Try to start a second instance, observe failure
Change-Id: Id02a831b249f6432b22741dbf5a5dc3919c3df03