Reduce expected length of VM's DICE chain
In preparation for only giving VMs the fragment of the DICE chain that
is unique to the VM, and so doesn't contain any device identifiers,
reduce the expected length of the DICE chain to the number of microdroid
boot stages. This will be tightened from a lower bound to an equality
once all the changes are complete.

Bug: 231422423
Test: TH
Change-Id: Ice82c7b6231d95c616c4bcaf7f1fddb9e759a7a8