compos: mount BuildManifestSystemExt.apk if exists
To specify the extra APK, we have to add two new VM config JSONs because
the APK may not exist, and there is no way to specify the config
dynamically.

Change VmParameters in order to derive the config path from two
variables at a more consistent place. As a result, there's no more
"default" config.

Bug: 246000387
Test: 1. manually add a jar from system_ext to SYSTEMSERVERCLASSPATH
      2. boot the VM, saw 2 directories in /mnt/extra-apk/
      (and with more local WIP changes, ComposHostTestCases passed)
Change-Id: I8aac49e0b347560fb65d8a8330e4333a29d90433