Restrict access to certain vm_payload APIs
Require the USE_CUSTOM_VIRTUAL_MACHINE permission in order to use
certain APIs from the VM payload that should not be exposed to all
clients of the AVF API. The permission is inferred from the use of a VM
config file, which requires the permission. The permission is only
granted to platform and test components.

Use this new ability to prevent VM payloads from accessing the raw DICE
chain and attestation CDI.

Fix: 243514248
Test: atest MicrodroidTests ComposHostTestCases
Change-Id: I1fd65ee1d0f624bc3ff9143f597e455c84ed2b02