pvmfw: Introduce pvmfw_embedded_key
pvmfw contains a reference public key against which it compares the key
used to sign the virtual machine payloads it receives. Export that
public key as pvmfw_embedded_key, a binary file intended to be built
into pvmfw. For release builds, the pvmfw image is expected to be
post-processed to find-and-replace (in-place) that key with the public
release key.

Bug: 255575669
Test: m pvmfw_embedded_key
Change-Id: Ia4490688cc3d1954a18e703a97812a9a28e63e7a