Statically link more libraries
In commit 11a80bef0c993b1f77771f1ff42e359976714ac5 we modified TestApk
so that the payload statically linked most libraries. But then we
added MicrodroidIdleNativeLib which didn't follow the same pattern.

So,
- Move MicrodroidIdleNativeLib under testapk.
- Statically link its libc++.
- Get rid of libiobenchmark which seemed redundant.

I haven't attempted to remove libbase.so and libc++.so from the
benchmark APK; the files are used both in host & VM, and I suspect
doing so would be a net loss.

Bug: 255920361
Test: Manual: Build, check .so files contained in the APKs
Test: atest MicrodroidTests MicrodroidBenchmarkApp
Change-Id: I6620f179536da39bab3a4ebc4f288af333548562