pvmfw: Receive payload location from DT
To unify the cases where pvmfw is loaded into pVM address space by pKVM
(the VMM is then allowed to set registers x1 and x2) or by the VMM
itself (all registers except x0 are silently zeroed by pKVM), locate the
pre-loaded kernel through the DT, in a similar way to the ramdisk.

Keep supporting the x1/x2 ABI when the legacy feature is enabled.

Bug: 253616035
Test: Boot pvmfw & check logs with a crosvm with crrev.com/c/4064759
Change-Id: I74d3e52025fa1fbc5e74246e9fb09a7fb46222a0