Add some unit tests
There are a bunch of getters and helper methods that we don't exercise
in our tests, so I'm adding some tests to fix that. These are "unit
tests" in a loose sense - no VM is ever created, but they do involve
running code in an APK, touching the filesystem etc.

Also rearranged the ordering of members in the classes so e.g. all the
instance variables are together, which makes it easier to check this.

Bug: 244561836
Test: atest MicrodroidTests
Change-Id: I6887b203c18a021eb2415f2624900012040310e1