Add microdroid_manager.init_done - set init done my MM
Add this sysprop - this is used by Microdroid manager to declare that
initalization is complete as far as MM is considered. Init triggers
actions such as stop ueventd on it.

Why not use dev.bootcomplete? - Because we don't want to set it before
post-fs-data is triggered (that starts authfs_service).

Bug: 260713790
Bug: 260005615
Test: atest MicrodroidTestApp

Change-Id: I070f63678fd266b8e98234a5d260f93e72de7e3e