Restrict exception constructors
Our exception type is public, to allow API clients to catch it. The
constructors do not need to be part of the API, and do not need to be
called by anything other than the VM API implementation, so don't mark
them public. This allows us to add or remove constructors in the
future.

Bug: 261037705
Bug: 261089785
Test: Everything still builds
Test: atest MicrodroidTests
Change-Id: I31656353a82e13a9d26ca10b9c35458b47909b90