pvmfw: Loop SMCCC_TRNG_RND on NO_ENTROPY
As the firmware is allowed by the specification to return NO_ENTROPY:

    The call is non-blocking, when the available entropy is less than N,
    the call must returns `NO_ENTROPY`.
    There are system designs where FW may limit the rate at which
    entropy is requested by a caller.
    To enable this entropy request rate limitation, the call is allowed
    to return `NO_ENTROPY` even if sufficient conditioned

Loop until we receive the requested entropy instead of throwing an error
(resulting in a pvmfw panic).

Note that this could cause a pVM to become unresponsive during boot if
the firmware never returns the requested entropy (a behaviour that seems
to be compliant with the spec).

Bug: 270841564
Test: atest MicrodroidHostTests
Change-Id: If2bb3fba0dcbded3772a41c9b7e6e7fb1180f225