Disable ComposTestCase on Cuttlefish
It is very flaky, with a high level of OOM errors both inside &
outside the VM. This doesn't seem to indicate a real problem, but
causes presubmit problems.

Also revert commit 08653b3cedd569a33ef0f2dbbe876d40b901ecc5 and move
the test back to our presubmit. If that is successful, we can then do
the same for other TEST_MAPPING files.

Bug: 264496291
Test: atest ComposHostTestCases
Change-Id: I2487f555c35ffdb5610efb67146922762ee170a4