Bind ignored results.
std::io::Error implements Drop in some configurations. A value that
implements Drop will be dropped immediately when not bound (and _ does
not bind). This is asymptomatic in this case but we want to deny this
behavior globally because it has very surprising behavior when the
"unused" value is a lock.

Bug: 281730184
Test: mm
Change-Id: I325f7bb1ca8eb74f56fbb3926e4e79e833a265a4