Rely on logs instead to port forwarding in test
adbConnectToMicrodroid is flaky, the port forwarding fails (although
infrequently). This is the only test that relies on port forwarding,
rest use the MicrodroidBuilder class or inspect logs.

Change the test to use logs to determine if microdroid was up and
running.

Ignore-AOSP-First: This is a bug fix in udc-dev, will merge upstream separately.
Test: Run atest #testBootSucceedsWhenNonProtectedVmStartsWithImagesSignedWithDifferentKey
Bug: 279838948
Change-Id: Idfbe9d992994b164d285f59ecbbb4dfddaf9a74d