Remove redundant unsafe markers
Config::new is now safe, since it uses LayoutVerified to verify size &
alignment of the slice it is given. So remove the unsafe marker on it
and its callers and the related safety comments.

The only unsafety is where we create the mutable slice from a pointer,
and that is appropriately annotated. (I've updated the safety comments
there slightly.)

Bug: 275693559
Test: Flash new pvmfw, start protected VM
Change-Id: If5f4f791f7986f149dfbae03c9823362d954a666