Skip changingDebuggability related tests on Non-pVM
These are boot tests (more precisely, they test that VM does not boot)
when debuggability is changed. Debuggability changes the initrd which is
verified by pvmfw. They dont make much sense in cases of non-protected
VM where pvmfw is not in boot seq.

Note: the test still pass for non-protected VM on devices without debug
policy, the reason for that is an extra apex 'adbd' is included in the
VM, which results in MicrodroidData 'mismatch' result in boot failure.
But that is not the point of the test.

Bug: 301014052
Test: atest #changingNonDebuggableVmDebuggableInvalidatesVmIdentity &
check that it is indeed skipped
Merged-In: Ic2818a8f513c778ad1f5217d3641398c8df1719d

Change-Id: Ic2818a8f513c778ad1f5217d3641398c8df1719d
(cherry picked from commit 163029a9b5bd002262cf7c12d00d65bed40a31c7)